+++ changelog.md
@@ -5,8 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Add shared functions file 
+- Update hide plugins and include hummingbird plugin for non-Pie users
+- Add Security Headers functionality
 
 ## [1.4.0] 2026-01-29 
+- Refactoring & adding Redirection module
 
 ## [1.3.3] 2024-09-09 
 - Fix: multisite staging issue with new is staging check